move URI claim from start to validate
The change-store write now happens at the tail of validate.Process, after
all checks (duplicate, merge, metadata) pass. Per-queue partition leasing
makes the read-then-claim sequence within validate race-free.

Trade-off: requests that fail validation no longer leave orphan claim rows
in the change store. start becomes simpler — it persists the request and
publishes; the change-store dependency is gone.

- start.go / start_test.go: drop changeStore field, dependency, and tests
- validate.go: add claimURIs() helper called after metadata fetch
- validate_test.go: dup-detection table now expects Create on the success
  path (.AnyTimes() covers both success and rejection cases)
- example/server/orchestrator/main.go: stop passing changeStore to start
