+++ orchestrator/controller/start/BUILD.bazel
@@ -10,7 +10,6 @@ go_library(
         "//core/request",
         "//entity",
         "//entity/queue",
-        "//extension/changestore",
         "//extension/storage",
         "@com_github_uber_go_tally_v4//:tally",
         "@org_uber_go_zap//:zap",
@@ -26,7 +25,6 @@ go_test(
         "//core/errs",
         "//entity",
         "//entity/queue",
-        "//extension/changestore/mock",
         "//extension/queue/mock",
         "//extension/storage",
         "//extension/storage/mock",
