+++ orchestrator/controller/start/start.go
@@ -18,28 +18,24 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"time"
 
 	"github.com/uber-go/tally/v4"
 	"github.com/uber/submitqueue/core/consumer"
 	corerequest "github.com/uber/submitqueue/core/request"
 	"github.com/uber/submitqueue/entity"
 	entityqueue "github.com/uber/submitqueue/entity/queue"
-	"github.com/uber/submitqueue/extension/changestore"
 	"github.com/uber/submitqueue/extension/storage"
 	"go.uber.org/zap"
 )
 
 // Controller handles start queue messages.
-// It consumes requests, persists them to the request store, claims their URIs in the change
-// store, and publishes to the validate stage. Both writes are idempotent on retries; the
-// duplicate-detection check itself is performed downstream by the validate controller.
+// It consumes requests, persists them to storage, and publishes to the validate stage.
+// Duplicate detection and URI claiming happen downstream in the validate controller.
 // Implements consumer.Controller.
 type Controller struct {
 	logger        *zap.SugaredLogger
 	metricsScope  tally.Scope
 	store         storage.Storage
-	changeStore   changestore.ChangeStore
 	registry      consumer.TopicRegistry
 	topicKey      consumer.TopicKey
 	consumerGroup string
@@ -53,7 +49,6 @@ func NewController(
 	logger *zap.SugaredLogger,
 	scope tally.Scope,
 	store storage.Storage,
-	changeStore changestore.ChangeStore,
 	registry consumer.TopicRegistry,
 	topicKey consumer.TopicKey,
 	consumerGroup string,
@@ -62,7 +57,6 @@ func NewController(
 		logger:        logger.Named("start_controller"),
 		metricsScope:  scope.SubScope("start_controller"),
 		store:         store,
-		changeStore:   changeStore,
 		registry:      registry,
 		topicKey:      topicKey,
 		consumerGroup: consumerGroup,
@@ -109,18 +103,10 @@ func (c *Controller) Process(ctx context.Context, delivery consumer.Delivery) er
 
 	// Persist request to storage. ErrAlreadyExists means a queue redelivery of the same
 	// request_id (an at-least-once retry of THIS message), not a cross-request collision.
-	// Cross-request URI overlap is detected downstream in the validate controller.
+	// Cross-request URI overlap is detected and claims are written downstream in validate.
 	if err := c.store.GetRequestStore().Create(ctx, request); err != nil && !errors.Is(err, storage.ErrAlreadyExists) {
 		c.metricsScope.Counter("storage_errors").Inc(1)
 		return fmt.Errorf("failed to create request: %w", err)
-	}
-
-	// Claim this request's URIs in the change store. INSERT IGNORE makes this idempotent
-	// on retries with the same request_id. The validate controller reads from this store
-	// to detect cross-request URI overlap.
-	if err := c.claimURIs(ctx, request); err != nil {
-		c.metricsScope.Counter("change_store_errors").Inc(1)
-		return fmt.Errorf("failed to claim URIs for request %s: %w", request.ID, err)
 	}
 
 	// Record the "new" status in the request log
@@ -146,29 +132,6 @@ func (c *Controller) Process(ctx context.Context, delivery consumer.Delivery) er
 	c.metricsScope.Counter("processed").Inc(1)
 
 	return nil // Success - message will be acked
-}
-
-// claimURIs persists one ChangeRecord per URI in the request. Each Create call is
-// independent; the change store's per-PK idempotency makes the loop safe under
-// queue redelivery (same (Queue, URI, RequestID) is a no-op on retry). Different
-// requests with overlapping URIs do NOT collide on insert; the validate controller
-// queries the change store to detect that overlap.
-func (c *Controller) claimURIs(ctx context.Context, request entity.Request) error {
-	now := time.Now().UnixMilli()
-	for _, uri := range request.Change.URIs {
-		record := entity.ChangeRecord{
-			URI:       uri,
-			RequestID: request.ID,
-			Queue:     request.Queue,
-			CreatedAt: now,
-			UpdatedAt: now,
-			Version:   1,
-		}
-		if err := c.changeStore.Create(ctx, record); err != nil {
-			return fmt.Errorf("failed to claim uri=%s for request %s: %w", uri, request.ID, err)
-		}
-	}
-	return nil
 }
 
 // publish publishes a request ID to the specified topic key.
